+++ ntex/src/util/mod.rs
@@ -3,6 +3,7 @@ pub mod counter;
 mod extensions;
 pub mod inflight;
 pub mod keepalive;
+pub mod sink;
 pub mod stream;
 pub mod time;
 pub mod timeout;