+++ ntex/CHANGES.md
@@ -1,5 +1,15 @@
 # Changes
 
+## [0.3.6] - 2021-03-06
+
+* http.client: Fix WsConnection::start() definition
+
+* http.client: Introduce WsConnection::start_default() method
+
+* web: TestServer::ws() returns WsConnection
+
+* util: Add `SinkService` service
+
 ## [0.3.5] - 2021-03-04
 
 * framed: add high/low watermark for read/write buffers